# perf: added indexes for efficient queries in a multi-tenant setup, issue #2370