--- a/internal/storage/postgres/migrations/20250912101325_add_multi_tenant_indexes.sql
+++ b/internal/storage/postgres/migrations/20250912101325_add_multi_tenant_indexes.sql
@@ -2,7 +2,9 @@
 -- +goose Up
 CREATE INDEX CONCURRENTLY IF NOT EXISTS idx_transactions_tenant_id ON transactions (tenant_id, id);
 CREATE INDEX CONCURRENTLY IF NOT EXISTS idx_schema_tenant_version ON schema_definitions (tenant_id, version);
+DROP INDEX CONCURRENTLY IF EXISTS idx_schema_version;
 
 -- +goose Down
+CREATE INDEX CONCURRENTLY IF NOT EXISTS idx_schema_version ON schema_definitions (version);
 DROP INDEX CONCURRENTLY IF EXISTS idx_transactions_tenant_id;
 DROP INDEX CONCURRENTLY IF EXISTS idx_schema_tenant_version;