Details are in issue #2370

Here are results without and with indexes:

```
explain (analyze on, buffers on)
select MAX(id) from transactions where tenant_id = 't2';
/*
 Result  (cost=1103.95..1103.96 rows=1 width=8) (actual time=14.476..14.478 rows=1 loops=1)
  Buffers: shared hit=1636
  InitPlan 1 (returns $0)
    ->  Limit  (cost=0.29..1103.95 rows=1 width=8) (actual time=14.469..14.471 rows=1 loops=1)
          Buffers: shared hit=1636
          ->  Index Scan Backward using pk_transaction on transactions  (cost=0.29..5518.59 rows=5 width=8) (actual time=14.467..14.467 rows=1 loops=1)
                Index Cond: (id IS NOT NULL)
                Filter: ((tenant_id)::text = 't2'::text)
                Rows Removed by Filter: 110298
                Buffers: shared hit=1636
Planning Time: 0.126 ms
Execution Time: 14.507 ms
 */

CREATE INDEX CONCURRENTLY IF NOT EXISTS idx_transactions_tenant_id ON transactions (tenant_id, id);

explain (analyze on, buffers on)
select MAX(id) from transactions where tenant_id = 't2';

/*
 Result  (cost=4.39..4.40 rows=1 width=8) (actual time=0.053..0.054 rows=1 loops=1)
  Buffers: shared hit=1 read=3
  InitPlan 1 (returns $0)
    ->  Limit  (cost=0.42..4.39 rows=1 width=8) (actual time=0.050..0.050 rows=1 loops=1)
          Buffers: shared hit=1 read=3
          ->  Index Only Scan Backward using idx_transactions_tenant_id on transactions  (cost=0.42..20.29 rows=5 width=8) (actual time=0.048..0.049 rows=1 loops=1)
                Index Cond: ((tenant_id = 't2'::text) AND (id IS NOT NULL))
                Heap Fetches: 0
                Buffers: shared hit=1 read=3
Planning:
  Buffers: shared hit=19 read=1
Planning Time: 0.405 ms
Execution Time: 0.077 ms
 */
```

```
explain (analyze on, buffers on)
select version from schema_definitions where tenant_id = 't12' order by version DESC limit 1;
/*
 Limit  (cost=20.65..20.65 rows=1 width=21) (actual time=0.134..0.135 rows=1 loops=1)
  Buffers: shared hit=63
  ->  Sort  (cost=20.65..20.67 rows=10 width=21) (actual time=0.132..0.133 rows=1 loops=1)
        Sort Key: version DESC
        Sort Method: top-N heapsort  Memory: 25kB
        Buffers: shared hit=63
        ->  Index Only Scan using pk_schema_definition on schema_definitions  (cost=0.43..20.60 rows=10 width=21) (actual time=0.062..0.084 rows=150 loops=1)
              Index Cond: (tenant_id = 't12'::text)
              Heap Fetches: 0
              Buffers: shared hit=63
Planning Time: 0.159 ms
Execution Time: 0.198 ms
 */

CREATE INDEX idx_schema_tenant_version ON schema_definitions (tenant_id, version);

explain (analyze on, buffers on)
select version from schema_definitions where tenant_id = 't12' order by version DESC limit 1;
/*
 Limit  (cost=0.43..0.68 rows=1 width=21) (actual time=0.048..0.049 rows=1 loops=1)
  Buffers: shared hit=4
  ->  Index Only Scan Backward using idx_schema_tenant_version on schema_definitions  (cost=0.43..4.73 rows=17 width=21) (actual time=0.046..0.046 rows=1 loops=1)
        Index Cond: (tenant_id = 't12'::text)
        Heap Fetches: 0
        Buffers: shared hit=4
Planning Time: 0.227 ms
Execution Time: 0.074 ms
 */
```

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

* **Chores**
  * Database migration improves query performance in multi-tenant environments, speeding transaction and schema-related operations. Designed for safe, zero-downtime rollout with rollback support. Users should see faster page loads and smoother navigation under high tenant loads. No configuration changes or action required.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->