@zhurs 2025-08-12
Yes, sure.

The query plan after removing the index `idx_schema_version`:
```
Limit  (cost=0.43..0.68 rows=1 width=21) (actual time=0.061..0.062 rows=1 loops=1)
  Buffers: shared hit=4
  ->  Index Only Scan Backward using idx_schema_tenant_version on schema_definitions  (cost=0.43..4.73 rows=17 width=21) (actual time=0.060..0.060 rows=1 loops=1)
        Index Cond: (tenant_id = 't12'::text)
        Heap Fetches: 0
        Buffers: shared hit=4
Planning Time: 0.165 ms
Execution Time: 0.086 ms
```

I’ve also checked that all queries include the predicate `tenant_id=?`, so the new index should cover all use cases.